requirements.txt
django
pytz
psycopg2
django-dotenv
django-debug-toolbar
django-import-export
django-filter
django-el-pagination
django-crispy-forms
django-modeltranslation
dj-database-url
gunicorn
whitenoise
django-contrib-comments
django-widget-tweaks
django-phonenumber-field
ipython
